add videos to tutorials created only by the user is succesful

+++ server/schemas/resolvers.js
@@ -1,70 +1,42 @@
-const { Profile, Category, Tutorial, Comment } = require('../models');
+const { Profile, Category, Tutorial, Comment } = require('../models'); // Remove Video import
 const { signToken } = require('../utils/auth');
 const { AuthenticationError } = require('apollo-server-express');
 
 const resolvers = {
   Query: {
-    profiles: async () => {
-      return Profile.find({});
-    },
-
-    profile: async (_, { _id }) => {
-      return Profile.findById(_id);
-    },
-
+    profiles: async () => Profile.find({}),
+    profile: async (_, { _id }) => Profile.findById(_id),
     me: async (parent, args, context) => {
       if (context.user) {
-        // populate  the tutorials created by the logged-in user
         return Profile.findOne({ _id: context.user._id }).populate({
           path: 'tutorials',
           populate: 'category'
         });
       }
-      throw AuthenticationError;
+      throw new AuthenticationError('Not authenticated');
     },
-    // added query for locating single tutorial by its id and populating author, category and comments associated with it -tb
-    tutorial: async (parent, { _id }) => {
-      return Tutorial.findById(_id).populate('author category comments');
-    },
-
-    tutorials: async () => {
-      return Tutorial.find({}).populate('author category comments');
-    },
-
-    categories: async () => {
-      return Category.find({});
-    },
-
-    comments: async () => {
-      return Comment.find({});
-    }
+    tutorial: async (parent, { _id }) => Tutorial.findById(_id).populate('author category comments videos'),
+    tutorials: async () => Tutorial.find({}).populate('author category comments videos'),
+    categories: async () => Category.find({}),
+    comments: async () => Comment.find({})
   },
 
   Profile: {
-    tutorials: async (parent) => {
-      return Tutorial.find({ author: parent._id }).populate('category comments');
-    },
-    comments: async (parent) => {
-      return Comment.find({ author: parent._id });
-    }
+    tutorials: async (parent) => Tutorial.find({ author: parent._id }).populate('category comments videos'),
+    comments: async (parent) => Comment.find({ author: parent._id })
   },
 
   Tutorial: {
-    comments: async (parent) => {
-      return Comment.find({ tutorial: parent._id });
-    }
+    comments: async (parent) => Comment.find({ tutorial: parent._id }),
+    videos: async (parent) => parent.videos // Assuming videos are already populated
   },
 
   Category: {
-    tutorials: async (parent) => {
-      return Tutorial.find({ category: parent._id });
-    }
+    tutorials: async (parent) => Tutorial.find({ category: parent._id })
   },
 
   Comment: {
-    author: async (parent) => {
-      return Profile.findById(parent.author);
-    }
+    author: async (parent) => Profile.findById(parent.author)
   },
 
   Mutation: {
@@ -87,75 +59,71 @@ const resolvers = {
         throw new Error('An error occurred during sign up. Please try again.');
       }
     },
+    
     login: async (parent, { email, password }) => {
       const profile = await Profile.findOne({ email });
-
       if (!profile) {
-        throw AuthenticationError;
+        throw new AuthenticationError('No profile with this email found.');
       }
 
       const correctPw = await profile.isCorrectPassword(password);
-
       if (!correctPw) {
-        throw AuthenticationError;
+        throw new AuthenticationError('Incorrect password.');
       }
 
       const token = signToken(profile);
       return { token, profile };
     },
 
-
     addTutorial: async (parent, { title, category, content }, context) => {
       if (context.user) {
-        
-          // Find the category by name
-          const categoryDoc = await Category.findOne({ name: category });
-          if (!categoryDoc) {
-            throw new Error('You need to add a category.');
-          }
-
-          // Create the new tutorial
-          const newTutorial = await Tutorial.create({
-            title,
-            author: context.user._id,
-            category: categoryDoc._id,
-            content
-          });
-
-          await Profile.findByIdAndUpdate(
-            context.user._id,
-            { $addToSet: { tutorials: newTutorial._id } },
-            { new: true, runValidators: true })
-
-          return Tutorial.findById(newTutorial._id).populate('author category');
-
+        const categoryDoc = await Category.findOne({ name: category });
+        if (!categoryDoc) {
+          throw new Error('You need to add a category.');
         }
-  throw new AuthenticationError('Not authenticated');
-},
 
-    //! If remove mutation does not work, please remove "context" throughout the code and use "profileId" instead. Instead of context.user.id, use profile._id as well  -tb
-    //! profileId and profile._id is verified to be working to remove comments
+        const newTutorial = await Tutorial.create({
+          title,
+          author: context.user._id,
+          category: categoryDoc._id,
+          content
+        });
+
+        await Profile.findByIdAndUpdate(
+          context.user._id,
+          { $addToSet: { tutorials: newTutorial._id } },
+          { new: true, runValidators: true }
+        );
+
+        return Tutorial.findById(newTutorial._id).populate('author category');
+      }
+      throw new AuthenticationError('Not authenticated');
+    },
 
     removeTutorial: async (parent, { _id }, context) => {
       if (context.user) {
         try {
           const findTutorial = await Tutorial.findById(_id);
 
-          await Tutorial.findByIdAndDelete(_id)
+          if (!findTutorial) {
+            throw new Error('Tutorial not found');
+          }
 
-          await Profile.findByIdAndUpdate(Tutorial.author,
-            { $pull: { tutorials: _id } })
+          await Tutorial.findByIdAndDelete(_id);
 
-          return findTutorial
+          await Profile.findByIdAndUpdate(
+            findTutorial.author,
+            { $pull: { tutorials: _id } },
+            { new: true, runValidators: true }
+          );
 
+          return findTutorial;
         } catch (error) {
+          console.error('Remove tutorial error:', error);
           throw new AuthenticationError('Not authenticated');
         }
       }
     },
-        
-    //! If remove mutation does not work, please remove "context" throughout the code and use "profileId" instead. Instead of context.user.id, use profile._id as well  -tb
-    //! profileId and profile._id is verified to be working to remove comments
 
     updateTutorial: async (parent, { _id, title, category, content }, context) => {
       if (context.user) {
@@ -164,13 +132,13 @@ const resolvers = {
           if (!categoryDoc) {
             throw new Error('Category not found');
           }
-    
+
           const updatedTutorial = await Tutorial.findByIdAndUpdate(
             _id,
             { title, content, category: categoryDoc._id },
             { new: true }
           ).populate('category');
-    
+
           return updatedTutorial;
         } catch (error) {
           console.error('Update tutorial error:', error);
@@ -178,59 +146,62 @@ const resolvers = {
         }
       }
     },
-    
 
-    // added profileId and removed context.user from code. Removed finding tutorial by id and instead replaced it with creating a comment and then finding and updating the tutorial. Then returning the mutation with the populated data
     addComment: async (parent, { profileId, tutorialId, content }, context) => {
       if (context.user) {
-      try {
+        try {
+          const addComment = await Comment.create({
+            content,
+            author: profileId,
+            tutorial: tutorialId
+          });
 
-        const addComment = await Comment.create({
-          content,
-          author: profileId,
-          tutorial: tutorialId
-        })
+          await Profile.findByIdAndUpdate(
+            profileId,
+            { $addToSet: { comments: addComment._id } },
+            { new: true, runValidators: true }
+          );
 
-        const updateProfile = await Profile.findByIdAndUpdate(
-          profileId._id,
-          { $addToSet: { comments: addComment._id } },
-          { new: true, runValidators: true }
-        )
+          await Tutorial.findByIdAndUpdate(
+            tutorialId,
+            { $addToSet: { comments: addComment._id } },
+            { new: true, runValidators: true }
+          );
 
-        const updateTutorial = await Tutorial.findByIdAndUpdate(
-          tutorialId,
-          { $addToSet: { comments: addComment._id } },
-          { new: true, runValidators: true })
-
-        return addComment.populate('author tutorial');
+          return addComment.populate('author tutorial');
         } catch (error) {
+          console.error('Add comment error:', error);
           throw new AuthenticationError('Not authenticated');
         }
       }
     },
 
-    //! If remove mutation does not work, please remove "context" throughout the code and use profileId instead -tb
-    //! profileId is verified to be working to remove comments
-
     removeComment: async (parent, { _id }, context) => {
       if (context.user) {
-      try {
-        const comment = await Comment.findById(_id);
+        try {
+          const comment = await Comment.findById(_id);
 
-        const deleteComment = await Comment.findByIdAndDelete(_id);
+          if (!comment) {
+            throw new Error('Comment not found');
+          }
 
-        const updateProfile = await Profile.findByIdAndUpdate(comment.author,
-          { $pull: { comments: _id } },
-          { new: true, runValidators: true })
+          await Comment.findByIdAndDelete(_id);
 
+          await Profile.findByIdAndUpdate(
+            comment.author,
+            { $pull: { comments: _id } },
+            { new: true, runValidators: true }
+          );
 
-        const updateTutorial = await Tutorial.findByIdAndUpdate(comment.tutorial,
-          { $pull: { comments: _id } },
-          { new: true, runValidators: true });
+          await Tutorial.findByIdAndUpdate(
+            comment.tutorial,
+            { $pull: { comments: _id } },
+            { new: true, runValidators: true }
+          );
 
-        return comment;
-
+          return comment;
         } catch (error) {
+          console.error('Remove comment error:', error);
           throw new AuthenticationError('Not authenticated');
         }
       }
@@ -238,54 +209,70 @@ const resolvers = {
 
     updateComment: async (parent, { _id, content }, context) => {
       if (context.user) {
-      try {
-        // find tutorial by id and update with new values
-        const updateComment = await Comment.findOneAndUpdate({ _id }, { content });
+        try {
+          const updateComment = await Comment.findByIdAndUpdate(
+            _id,
+            { content },
+            { new: true }
+          );
 
-        return updateComment;
-
+          return updateComment;
         } catch (error) {
+          console.error('Update comment error:', error);
           throw new AuthenticationError('Not authenticated');
         }
       }
     },
 
-    // Mutation to save a video as a tutorial
-    saveVideoToTutorial: async (parent, { title, videoId, thumbnail, content, c }, context) => {
+    saveVideoToTutorial: async (parent, { title, videoId, thumbnail, content, tutorialId }, context) => {
       if (!context.user) {
         throw new AuthenticationError('You need to be logged in!');
       }
 
       try {
-        if (!videoId || !thumbnail  ) {
-          throw new Error('videoId, thumbnail, and category are required');
+        console.log('Saving video with parameters:', { title, videoId, thumbnail, content, tutorialId });
+
+        if (!videoId || !thumbnail) {
+          console.error('Missing required fields:', { videoId, thumbnail });
+          throw new Error('videoId and thumbnail are required');
         }
 
-        const tutorial = await Tutorial.create({
+        const tutorial = await Tutorial.findById(tutorialId);
+        if (!tutorial) {
+          console.error('Tutorial not found for id:', tutorialId);
+          throw new Error('Tutorial not found');
+        }
+
+        console.log('Tutorial before update:', tutorial);
+
+        const video = {
           title,
           videoId,
           thumbnail,
           content,
-          author: context.user._id,
-        });
+        };
 
-        const populatedTutorial = await Tutorial.findById(tutorial._id).populate('category').exec();
-        await Profile.findByIdAndUpdate(
-          context.user._id,
-          { $addToSet: { tutorials: tutorial._id } },
-          { new: true }
-        );
+        console.log('Video object:', video);
+
+        tutorial.videos.push(video);
+
+        await tutorial.save();
+
+        const populatedTutorial = await Tutorial.findById(tutorialId)
+          .populate('author')
+          .populate('category')
+          .populate('videos')
+          .exec();
+
+        console.log('Tutorial after update:', populatedTutorial);
 
         return populatedTutorial;
       } catch (error) {
+        console.error('Error saving video to tutorial:', error);
         throw new Error('Error saving video to tutorial: ' + error.message);
       }
     },
   },
 };
-
-
-
-
 
 module.exports = resolvers;
